==> code files/pilot/3_descriptive_stats.Rmd <==
---
title: "3_descriptive_stats"
output: html_notebook
---
# Library
```{r}
library(tidyverse)
`%notin%`<-Negate(`%in%`)
```

# Read Data
```{r}
# Import raw data
data_raw <- read_csv(here(
  "code files",
  "pilot",
  "data",
  "survey_data.csv"
))

choice_data <- read_csv(here(
  "code files",
  "old code",
  "testing_initial_data_modeling",
  "battery_choice_data.csv"
))

data_full<-data_raw %>% 
  filter(session_id %in% choice_data$session_id) %>% 
  select(-starts_with("time_"))
```


```{r}

```
